rename instanceMnaager to manager
